requirements.txt
pyspark==3.5.0
pandas==2.1.4
matplotlib==3.8.2
tabulate==0.9.0
colorama==0.4.6
python-dateutil==2.8.2
